requirements.txt
fastapi 
uvicorn 
pillow 
google-cloud-aiplatform



